+++ StripePayments/StripePayments/Source/API Bindings/Models/PaymentMethods/STPPaymentMethodEnums.swift
@@ -108,6 +108,8 @@ import Foundation
     case naverPay
     /// A PAYCO payment method
     case payco
+    /// A SeQura payment method
+    case sequra
     /// An unknown type.
     case unknown
 
@@ -214,6 +216,8 @@ import Foundation
             return "Naver Pay"
         case .payco:
             return "PAYCO"
+        case .sequra:
+            return "SeQura"
         case .cardPresent,
             .unknown:
             return STPLocalizedString("Unknown", "Default missing source type label")
@@ -323,6 +327,8 @@ import Foundation
             return "naver_pay"
         case .payco:
             return "payco"
+        case .sequra:
+            return "sequra"
         }
     }