+++ CHANGELOG.md
@@ -4,6 +4,7 @@ PATCH
 ## X.Y.Z - changes pending release
 
 ### Payments
+* [Added] Added API bindings for SeQura.
 * [Added] Added API bindings for Korean cards.
 * [Added] Added API bindings for Naver Pay.
 * [Added] Added API bindings for PAYCO.